fix: consolidate duplicate CPU type definitions
- Removed duplicate CPU6502State from core/types/cpu.ts
- Moved all CPU types to core/cpu6502/types.ts including:
  - DisassemblyLine, TraceEntry, CPU6502DebugExtensions
  - CPU6502WithDebug type helper
- Updated core/types/index.ts to re-export from cpu6502 module
- All existing imports continue to work through re-exports
