+++ src/core/cpu6502/types.ts
@@ -3,6 +3,45 @@
  */
 
 import type Bus from '../Bus';
+import type { StateBase } from '../types/state';
+
+/**
+ * Complete state representation of the 6502 CPU for serialization
+ */
+export interface CPU6502State extends StateBase {
+    // Registers
+    PC: number;  // Program Counter
+    A: number;   // Accumulator
+    X: number;   // X Index Register
+    Y: number;   // Y Index Register
+    S: number;   // Stack Pointer
+    
+    // Status flags
+    N: number;   // Negative
+    Z: number;   // Zero
+    C: number;   // Carry
+    V: number;   // Overflow
+    I: number;   // Interrupt Disable
+    D: number;   // Decimal Mode
+    
+    // Interrupt lines
+    irq: number;
+    nmi: number;
+    
+    // Execution state
+    cycles: number;
+    opcode: number;
+    address: number;
+    data: number;
+    
+    // Interrupt handling state
+    pendingIrq: number;
+    pendingNmi: number;
+    
+    // Cycle-accurate timing state (optional)
+    cycleAccurateMode?: boolean;
+    currentInstructionCycles?: number;
+}
 
 /**
  * Interface for CPU6502 class that defines all methods used by opcodes and other modules
@@ -142,3 +181,40 @@ export interface CPU6502Interface {
     pendingIrq?: number;
     pendingNmi?: number;
 }
+
+/**
+ * Disassembly result for a single instruction
+ */
+export interface DisassemblyLine {
+    address: number;
+    bytes: number[];
+    mnemonic: string;
+    operand?: string;
+}
+
+/**
+ * Instruction trace entry for debugging
+ */
+export interface TraceEntry {
+    pc: number;
+    opcode: number;
+    mnemonic: string;
+    operand?: string;
+    cycles: number;
+    timestamp?: number;
+}
+
+/**
+ * Optional debug extensions for CPU6502
+ */
+export interface CPU6502DebugExtensions {
+    /** Disassemble instructions starting from given PC */
+    disassemble?: (pc: number, count: number) => DisassemblyLine[];
+    /** Recent instruction execution trace */
+    trace?: TraceEntry[];
+}
+
+/**
+ * CPU6502 with optional debug extensions
+ */
+export type CPU6502WithDebug<T> = T & Partial<CPU6502DebugExtensions>;
